Add class User
