+++ src/com/thunder_cut/netio/Server.java
@@ -175,12 +175,12 @@ public class Server implements ConnectionCallback {
             String command = new String(parsed.getData(), StandardCharsets.UTF_8);
             String[] args = command.split(" ");
             if (CommandType.getCommand(args[0]) == CommandType.NAME) {
-                source.setName(command.substring(command.indexOf(' ') + 1));
+                source.getUser().setName(command.substring(command.indexOf(' ') + 1));
                 send(new Data(DataType.LIST, 0, connectionsToString().getBytes(StandardCharsets.UTF_8)).toEncrypted(secretKey));
             }
         } else {
             if (parsed.dataType == DataType.MESSAGE) {
-                System.out.println(source.getName() + " (" + source.id + "): " + new String(parsed.getData(), StandardCharsets.UTF_8));
+                System.out.println(source.getUser().getName() + " (" + source.id + "): " + new String(parsed.getData(), StandardCharsets.UTF_8));
             }
             parsed.setSrcId(source.id);
             send(parsed.toEncrypted(secretKey));
@@ -202,7 +202,7 @@ public class Server implements ConnectionCallback {
         for (Connection connection : array) {
             stringBuilder.append(connection.id);
             stringBuilder.append('/');
-            stringBuilder.append(connection.getName());
+            stringBuilder.append(connection.getUser().getName());
             stringBuilder.append('/');
         }
         stringBuilder.deleteCharAt(stringBuilder.length() - 1);
